Correction norme REST

+++ src/main/java/com/loicmaria/api/controller/AuthorController.java
@@ -4,6 +4,7 @@ package com.loicmaria.api.controller;
 import com.loicmaria.api.DTO.AuthorDto;
 import com.loicmaria.api.service.AuthorServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,61 +18,55 @@ public class AuthorController {
     @Autowired
     AuthorServiceImpl authorService;
 
+    //      CRUD Operations
+    //----------------------------------------------------------------------------------------------------------------
 
-    /**
-     * Create - Add a new author
-     *
-     * @param authorDto An object author
-     * @return ResponseEntity.ok
-     */
-    @PostMapping("/create")
+    @PostMapping
     public ResponseEntity<?> createAuthor(@RequestBody AuthorDto authorDto) {
-        return ResponseEntity.ok(authorService.save(authorDto));
+        AuthorDto createdAuthor = authorService.save(authorDto);
+
+        if (createdAuthor != null) {
+            return ResponseEntity
+                    .status(HttpStatus.CREATED)
+                    .body(createdAuthor);
+        } else {
+            return ResponseEntity
+                    .status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body("Erreur lors de la création de l'auteur.");
+        }
     }
 
-
-    /**
-     * Read - Get one author
-     *
-     * @param id The id of the author
-     * @return An Author object full filled
-     */
-    @GetMapping("/id/{id}")
-    public AuthorDto getAuthor(@PathVariable("id") int id) {
-        AuthorDto authorDto = authorService.get(id);
-        return authorDto;
+    @GetMapping("/{id}")
+    public AuthorDto getAuthor(@PathVariable int id) {
+        return authorService.get(id);
     }
 
-    /**
-     * Read - Get all authors
-     *
-     * @return - An Iterable object of Author full filled
-     */
-    @GetMapping("/all")
+    @GetMapping
     public Collection<AuthorDto> getAuthors() {
         return authorService.getter();
     }
 
-    /**
-     * Update - Update an existing author
-     *
-     * @param authorDto - The author object updated
-     * @return The currentAuthor if he is present or null
-     */
-    @PutMapping("/update")
-    public AuthorDto updateAuthor(@RequestBody AuthorDto authorDto) {
+    @PutMapping("/{id}")
+    public ResponseEntity<AuthorDto> updateAuthor(
+            @PathVariable int id,
+            @RequestBody AuthorDto authorDto) {
+        
+        if (id != authorDto.getId()) {
+            return ResponseEntity.badRequest().build();
+        }
         authorService.save(authorDto);
-        return authorDto;
+        return ResponseEntity.ok(authorDto);
     }
+    
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteAuthor(@PathVariable int id) {
+        boolean authorExists = authorService.exists(id);
 
-
-    /**
-     * Delete - Delete an author
-     *
-     * @param id - The id of the author to delete
-     */
-    @DeleteMapping("/id/{id}")
-    public void deleteAuthor(@PathVariable("id") int id) {
-        authorService.delete(id);
+        if (authorExists){
+            authorService.delete(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
